requirements.txt
fastapi>=0.99.1
pydantic>=1.8.1
SQLAlchemy>=2.0.17
passlib>=1.7.4
alembic>=1.11.1
uvicorn>=0.22.0
psycopg2>=2.9.6
kafka-python>=2.0.2
aiokafka>=0.8.1
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
pyjwt>=2.7.0